requirements_with_opt.txt
altgraph==0.17
appdirs==1.4.4
astor==0.8.1
astroid==2.3.2
autopep8==1.5.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.1
decorator==4.4.2
distlib==0.3.2
filelock==3.0.12
future==0.18.2
gast==0.3.3
idna==2.10
importlib-metadata==4.6.0
isort==4.3.21
joblib==1.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.19.3
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
paddlepaddle==2.1.1
pefile==2021.5.24
Pillow==8.2.0
pipenv==2021.5.29
protobuf==3.17.3
pycodestyle==2.6.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.2
pylint==2.4.3
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pywin32-ctypes==0.2.0
PyYAML==5.4.1
QPT==1.0a9
requests==2.25.1
rope==0.18.0
scikit-learn==0.22.2
scipy==1.7.0
six==1.16.0
threadpoolctl==2.1.0
toml==0.10.1
typed-ast==1.4.0
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.4.7
virtualenv-clone==0.5.4
wget==3.2
wincertstore==0.2
wrapt==1.11.2
zipp==3.4.1
